backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
qdrant-client
psycopg2-binary
langgraph
httpx
sentence-transformers

requests
pandas
diskcache
yfinance
lxml
numpy
earningscall
python-dateutil
